Update HeatTrigger.py
This code is useful , but incredibly buggy, not terribly useful and the fact that I can no longer install PySpin on my machine makes it obsolete and unusable. 
The Neural network code remains functional, but this script is useless without PySpin and is unhelpful on linux as 
- MAC OSX runs PySpin
and 
- Ubuntu runs PyCapture2

both are too different to adjust at this point. The safest method is to use the Spinnaker_QT application to capture images. It contains both the serial number and temperature.