Some informative messages
Instead of the camera just taking images if it is above the temperature specified (the camera being able to take images at 90 when the specified temperature is 25), which is unhelpful for research purposes and for understanding how thermal noise affects images, the camera now will ONLY take images if the temperature of the camera is within 5 degrees Celsius of the specified temperature. Otherwise, the camera will be de-initialized and will tell the user to allow the camera to cool.